+++ lib/concurrent/actor.rb
@@ -1,6 +1,6 @@
 require 'concurrent/configuration'
 require 'concurrent/executor/serialized_execution'
-require 'concurrent/logging'
+require 'concurrent/concern/logging'
 require 'concurrent/synchronization_object'
 require 'concurrent/edge/future'
