+++ lib/concurrent/actor/behaviour/executes_context.rb
@@ -9,7 +9,7 @@ module Concurrent
 
         def on_event(event)
           context.on_event(event)
-          core.log Logging::DEBUG, "event: #{event.inspect}"
+          core.log DEBUG, "event: #{event.inspect}"
           super event
         end
       end
