+++ lib/concurrent/promise.rb
@@ -1,7 +1,6 @@
 require 'thread'
 require 'concurrent/errors'
 require 'concurrent/ivar'
-require 'concurrent/obligation'
 require 'concurrent/executor/executor'
 
 module Concurrent
