+++ spec/concurrent/agent_spec.rb
@@ -1,5 +1,5 @@
-require_relative 'dereferenceable_shared'
-require_relative 'observable_shared'
+require_relative 'concern/dereferenceable_shared'
+require_relative 'concern/observable_shared'
 
 module Concurrent
