+++ spec/concurrent/concern/dereferenceable_shared.rb
