+++ spec/concurrent/concern/obligation_shared.rb
