+++ spec/concurrent/concern/observable_shared.rb
